+++ src/Contracts/GatewayManager.php
@@ -2,6 +2,8 @@
 
 namespace DoubleThreeDigital\SimpleCommerce\Contracts;
 
+use Illuminate\Http\Request;
+
 interface GatewayManager
 {
     public function use($className): self;
@@ -19,4 +21,14 @@ interface GatewayManager
     public function getCharge($order);
 
     public function refundCharge($order);
+
+    public function callback(Request $request);
+
+    public function webhook(Request $request);
+
+    public function isOffsiteGateway(): bool;
+
+    public function withRedirectUrl(string $redirectUrl): self;
+
+    public function withErrorRedirectUrl(string $errorRedirectUrl): self;
 }