+++ src/Gateways/Builtin/PayPalGateway.php
@@ -5,6 +5,7 @@ namespace DoubleThreeDigital\SimpleCommerce\Gateways\Builtin;
 use DoubleThreeDigital\SimpleCommerce\Contracts\Gateway;
 use DoubleThreeDigital\SimpleCommerce\Contracts\Order;
 use DoubleThreeDigital\SimpleCommerce\Events\PostCheckout;
+use DoubleThreeDigital\SimpleCommerce\Exceptions\CustomerNotFound;
 use DoubleThreeDigital\SimpleCommerce\Exceptions\GatewayDoesNotSupportPurchase;
 use DoubleThreeDigital\SimpleCommerce\Exceptions\PayPalDetailsMissingOnOrderException;
 use DoubleThreeDigital\SimpleCommerce\Facades\Currency;
@@ -36,7 +37,7 @@ class PayPalGateway extends BaseGateway implements Gateway
 
     public function isOffsiteGateway(): bool
     {
-        return true;
+        return $this->config()->get('mode', 'offsite') === 'offsite';
     }
 
     public function prepare(Prepare $data): Response
@@ -78,26 +79,37 @@ class PayPalGateway extends BaseGateway implements Gateway
 
         return new Response(true, [
             'result' => [
-                'id' => $response->result->id,
+                'id'            => $response->result->id,
+                'currency_code' => $response->result->purchase_units[0]->amount->currency_code,
             ],
-        ], $checkoutUrl->href);
+        ], $this->isOffsiteGateway() ? $checkoutUrl->href : '');
     }
 
     public function purchase(Purchase $data): Response
     {
-        // We don't actually do anything here as PayPal is an
-        // off-site gateway, so it has it's own checkout page.
+        if ($this->isOffsiteGateway()) {
+            throw new GatewayDoesNotSupportPurchase("Gateway [paypal] does not support the `purchase` method.");
+        }
 
-        throw new GatewayDoesNotSupportPurchase("Gateway [paypal] does not support the `purchase` method.");
+        $this->setupPayPal();
+
+        $request = new OrdersGetRequest($data->request()->payment_id);
+
+        /** @var \PayPalHttp\HttpResponse $response */
+        $response = $this->paypalClient->execute($request);
+
+        return new Response($response->result->status === 'APPROVED');
     }
 
     public function purchaseRules(): array
     {
-        // PayPal is off-site, therefore doesn't use the traditional
-        // checkout process provided by Simple Commerce. Hence why no rules
-        // are defined here.
+        if ($this->isOffsiteGateway()) {
+            return [];
+        }
 
-        return [];
+        return [
+            'payment_id' => 'required|string',
+        ];
     }
 
     public function getCharge(Order $order): Response
@@ -173,10 +185,10 @@ class PayPalGateway extends BaseGateway implements Gateway
             $response = $this->paypalClient->execute($request);
             $responseBody = json_decode(json_encode($response->result), true);
 
-            if (! $order->customer() && $responseBody['payer']['name'] || $responseBody['payer']['email_address']) {
-                $customer = Customer::findByEmail($responseBody['payer']['email_address']);
-
-                if (! $customer) {
+            if (is_null($order->customer()) && $responseBody['payer']['name'] && $responseBody['payer']['email_address']) {
+                try {
+                    $customer = Customer::findByEmail($responseBody['payer']['email_address']);
+                } catch (CustomerNotFound $e) {
                     $customer = Customer::create([
                         'name' => $responseBody['payer']['name']['given_name'] . ' ' . $responseBody['payer']['name']['surname'],
                         'email' => $responseBody['payer']['email_address'],