+++ src/Orders/Calculator.php
@@ -66,9 +66,15 @@ class Calculator implements Contract
         $product = ProductAPI::find($lineItem['product']);
 
         if ($product->purchasableType() === 'variants') {
-            $productPrice = $product->variant(
+            $variant = $product->variant(
                 isset($lineItem['variant']['variant']) ? $lineItem['variant']['variant'] : $lineItem['variant']
-            )->price();
+            );
+
+            if (SimpleCommerce::$productVariantPriceHook) {
+                $productPrice = (SimpleCommerce::$productVariantPriceHook)($this->order, $product, $variant);
+            } else {
+                $productPrice = $variant->price();
+            }
 
             // Ensure we strip any decimals from price
             $productPrice = (int) str_replace('.', '', (string) $productPrice);
@@ -81,7 +87,11 @@ class Calculator implements Contract
             ];
         }
 
-        $productPrice = $product->get('price');
+        if (SimpleCommerce::$productPriceHook) {
+            $productPrice = (SimpleCommerce::$productPriceHook)($this->order, $product);
+        } else {
+            $productPrice = $product->get('price');
+        }
 
         // Ensure we strip any decimals from price
         $productPrice = (int) str_replace('.', '', (string) $productPrice);