+++ tests/Http/Controllers/CheckoutControllerTest.php
@@ -845,6 +845,396 @@ class CheckoutControllerTest extends TestCase
     }
 
     /** @test */
+    public function cant_post_checkout_when_product_has_a_single_item_left_in_stock_and_single_quantity_in_cart()
+    {
+        Event::fake();
+
+        $product = Product::create([
+            'title' => 'Bacon',
+            'price' => 5000,
+            'stock' => 1,
+        ]);
+
+        $order = Order::create([
+            'items' => [
+                [
+                    'id'       => Stache::generateId(),
+                    'product'  => $product->id,
+                    'quantity' => 1,
+                    'total'    => 5000,
+                ],
+            ],
+            'grand_total' => 5000,
+        ]);
+
+        $this
+            ->withSession(['simple-commerce-cart' => $order->id])
+            ->post(route('statamic.simple-commerce.checkout.store'), [
+                'name'         => 'Smelly Joe',
+                'email'        => 'smelly.joe@example.com',
+                'gateway'      => DummyGateway::class,
+                'card_number'  => '4242424242424242',
+                'expiry_month' => '01',
+                'expiry_year'  => '2025',
+                'cvc'          => '123',
+            ])
+            ->assertRedirect();
+
+        $order->fresh();
+
+        // Assert events have been dispatched
+        Event::assertDispatched(PreCheckout::class);
+        Event::assertDispatched(PostCheckout::class);
+
+        // Assert order has been marked as paid
+        $this->assertTrue($order->published);
+
+        $this->assertTrue($order->get('is_paid'));
+        $this->assertNotNull($order->get('paid_date'));
+
+        // Assert stock has been reduced
+        $product->fresh();
+        $this->assertSame($product->get('stock'), 0);
+
+        Event::assertDispatched(StockRunningLow::class);
+        Event::assertNotDispatched(StockRunOut::class);
+
+        // Finally, assert order is no longer attached to the users' session
+        $this->assertFalse(session()->has('simple-commerce-cart'));
+    }
+
+    /** @test */
+    public function can_post_checkout_with_variant_product_with_stock_counter()
+    {
+        Event::fake();
+
+        $product = Product::create([
+            'title' => 'Bacon',
+            'product_variants' => [
+                'variants' => [
+                    [
+                        'name'   => 'Colours',
+                        'values' => [
+                            'Red',
+                        ],
+                    ],
+                    [
+                        'name'   => 'Sizes',
+                        'values' => [
+                            'Small',
+                        ],
+                    ],
+                ],
+                'options' => [
+                    [
+                        'key'     => 'Red_Small',
+                        'variant' => 'Red Small',
+                        'price'   => 5000,
+                        'stock'   => 50,
+                    ],
+                ],
+            ],
+        ]);
+
+        $order = Order::create([
+            'items' => [
+                [
+                    'id'       => Stache::generateId(),
+                    'product'  => $product->id,
+                    'variant'  => 'Red_Small',
+                    'quantity' => 1,
+                    'total'    => 5000,
+                ],
+            ],
+            'grand_total' => 5000,
+        ]);
+
+        $this
+            ->withSession(['simple-commerce-cart' => $order->id])
+            ->post(route('statamic.simple-commerce.checkout.store'), [
+                'name'         => 'Smelly Joe',
+                'email'        => 'smelly.joe@example.com',
+                'gateway'      => DummyGateway::class,
+                'card_number'  => '4242424242424242',
+                'expiry_month' => '01',
+                'expiry_year'  => '2025',
+                'cvc'          => '123',
+            ]);
+
+        $order->fresh();
+
+        // Assert events have been dispatched
+        Event::assertDispatched(PreCheckout::class);
+        Event::assertDispatched(PostCheckout::class);
+
+        // Assert order has been marked as paid
+        $this->assertTrue($order->published);
+
+        $this->assertTrue($order->get('is_paid'));
+        $this->assertNotNull($order->get('paid_date'));
+
+        // Assert stock has been reduced
+        $product->fresh();
+
+        $this->assertSame($product->variant('Red_Small')->stockCount(), 49);
+
+        // Finally, assert order is no longer attached to the users' session
+        $this->assertFalse(session()->has('simple-commerce-cart'));
+    }
+
+    /** @test */
+    public function can_post_checkout_when_variant_product_is_running_low_on_stock()
+    {
+        Config::set('simple-commerce.low_stock_threshold', 10);
+
+        Event::fake();
+
+        $product = Product::create([
+            'title' => 'Bacon',
+            'product_variants' => [
+                'variants' => [
+                    [
+                        'name'   => 'Colours',
+                        'values' => [
+                            'Red',
+                        ],
+                    ],
+                    [
+                        'name'   => 'Sizes',
+                        'values' => [
+                            'Small',
+                        ],
+                    ],
+                ],
+                'options' => [
+                    [
+                        'key'     => 'Red_Small',
+                        'variant' => 'Red Small',
+                        'price'   => 5000,
+                        'stock'   => 9,
+                    ],
+                ],
+            ],
+        ]);
+
+        $order = Order::create([
+            'items' => [
+                [
+                    'id'       => Stache::generateId(),
+                    'product'  => $product->id,
+                    'variant'  => 'Red_Small',
+                    'quantity' => 1,
+                    'total'    => 5000,
+                ],
+            ],
+            'grand_total' => 5000,
+        ]);
+
+        $this
+            ->withSession(['simple-commerce-cart' => $order->id])
+            ->post(route('statamic.simple-commerce.checkout.store'), [
+                'name'         => 'Smelly Joe',
+                'email'        => 'smelly.joe@example.com',
+                'gateway'      => DummyGateway::class,
+                'card_number'  => '4242424242424242',
+                'expiry_month' => '01',
+                'expiry_year'  => '2025',
+                'cvc'          => '123',
+            ]);
+
+        $order->fresh();
+
+        // Assert events have been dispatched
+        Event::assertDispatched(PreCheckout::class);
+        Event::assertDispatched(PostCheckout::class);
+
+        // Assert order has been marked as paid
+        $this->assertTrue($order->published);
+
+        $this->assertTrue($order->get('is_paid'));
+        $this->assertNotNull($order->get('paid_date'));
+
+        // Assert stock has been reduced
+        $product->fresh();
+        $this->assertSame($product->variant('Red_Small')->stockCount(), 8);
+
+        Event::assertDispatched(StockRunningLow::class);
+
+        // Finally, assert order is no longer attached to the users' session
+        $this->assertFalse(session()->has('simple-commerce-cart'));
+    }
+
+    /** @test */
+    public function cant_post_checkout_when_variant_product_has_no_stock()
+    {
+        Event::fake();
+
+        $product = Product::create([
+            'title' => 'Bacon',
+            'product_variants' => [
+                'variants' => [
+                    [
+                        'name'   => 'Colours',
+                        'values' => [
+                            'Red',
+                        ],
+                    ],
+                    [
+                        'name'   => 'Sizes',
+                        'values' => [
+                            'Small',
+                        ],
+                    ],
+                ],
+                'options' => [
+                    [
+                        'key'     => 'Red_Small',
+                        'variant' => 'Red Small',
+                        'price'   => 5000,
+                        'stock'   => 0,
+                    ],
+                ],
+            ],
+        ]);
+
+        $order = Order::create([
+            'items' => [
+                [
+                    'id'       => Stache::generateId(),
+                    'product'  => $product->id,
+                    'variant'  => 'Red_Small',
+                    'quantity' => 1,
+                    'total'    => 5000,
+                ],
+            ],
+            'grand_total' => 5000,
+        ]);
+
+        $this
+            ->withSession(['simple-commerce-cart' => $order->id])
+            ->post(route('statamic.simple-commerce.checkout.store'), [
+                'name'         => 'Smelly Joe',
+                'email'        => 'smelly.joe@example.com',
+                'gateway'      => DummyGateway::class,
+                'card_number'  => '4242424242424242',
+                'expiry_month' => '01',
+                'expiry_year'  => '2025',
+                'cvc'          => '123',
+            ])
+            ->assertRedirect()
+            ->assertSessionHasErrors();
+
+        $order->fresh();
+
+        // Assert the line item has been wiped out
+        $this->assertSame($order->lineItems()->count(), 0);
+        $this->assertSame($order->get('grand_total'), 0);
+
+        // Assert events have been dispatched
+        Event::assertDispatched(PreCheckout::class);
+        Event::assertNotDispatched(PostCheckout::class);
+
+        // Assert order has been marked as paid
+        $this->assertFalse($order->published);
+
+        $this->assertFalse($order->get('is_paid'));
+        $this->assertNull($order->get('paid_date'));
+
+        // Assert stock has been reduced
+        $product->fresh();
+        $this->assertSame($product->variant('Red_Small')->stockCount(), 0);
+
+        Event::assertNotDispatched(StockRunningLow::class);
+        Event::assertDispatched(StockRunOut::class);
+
+        // Finally, assert order is no longer attached to the users' session
+        $this->assertTrue(session()->has('simple-commerce-cart'));
+    }
+
+    /** @test */
+    public function can_post_checkout_when_variant_product_has_a_single_item_left_in_stock_and_single_quantity_in_cart()
+    {
+        Config::set('simple-commerce.low_stock_threshold', 10);
+
+        Event::fake();
+
+        $product = Product::create([
+            'title' => 'Bacon',
+            'product_variants' => [
+                'variants' => [
+                    [
+                        'name'   => 'Colours',
+                        'values' => [
+                            'Red',
+                        ],
+                    ],
+                    [
+                        'name'   => 'Sizes',
+                        'values' => [
+                            'Small',
+                        ],
+                    ],
+                ],
+                'options' => [
+                    [
+                        'key'     => 'Red_Small',
+                        'variant' => 'Red Small',
+                        'price'   => 5000,
+                        'stock'   => 1,
+                    ],
+                ],
+            ],
+        ]);
+
+        $order = Order::create([
+            'items' => [
+                [
+                    'id'       => Stache::generateId(),
+                    'product'  => $product->id,
+                    'variant'  => 'Red_Small',
+                    'quantity' => 1,
+                    'total'    => 5000,
+                ],
+            ],
+            'grand_total' => 5000,
+        ]);
+
+        $this
+            ->withSession(['simple-commerce-cart' => $order->id])
+            ->post(route('statamic.simple-commerce.checkout.store'), [
+                'name'         => 'Smelly Joe',
+                'email'        => 'smelly.joe@example.com',
+                'gateway'      => DummyGateway::class,
+                'card_number'  => '4242424242424242',
+                'expiry_month' => '01',
+                'expiry_year'  => '2025',
+                'cvc'          => '123',
+            ]);
+
+        $order->fresh();
+
+        // Assert events have been dispatched
+        Event::assertDispatched(PreCheckout::class);
+        Event::assertDispatched(PostCheckout::class);
+
+        // Assert order has been marked as paid
+        $this->assertTrue($order->published);
+
+        $this->assertTrue($order->get('is_paid'));
+        $this->assertNotNull($order->get('paid_date'));
+
+        // Assert stock has been reduced
+        $product->fresh();
+        $this->assertSame($product->variant('Red_Small')->stockCount(), 0);
+
+        Event::assertDispatched(StockRunningLow::class);
+        Event::assertNotDispatched(StockRunOut::class);
+
+        // Finally, assert order is no longer attached to the users' session
+        $this->assertFalse(session()->has('simple-commerce-cart'));
+    }
+
+    /** @test */
     public function can_post_checkout_and_ensure_remaining_request_data_is_saved_to_order()
     {
         Event::fake();